refactor: reorganized package structure
