+++ src/reeruryu/week1/P12909.java
@@ -1,4 +1,4 @@
-package reeruryu;
+package reeruryu.week1;
 
 /*
 문제: 프로그래머스 - 올바른 괄호 / 난이도: Level 2
