+++ src/reeruryu/week1/P42576.java
@@ -1,4 +1,4 @@
-package reeruryu;
+package reeruryu.week1;
 
 /*
 문제: 프로그래머스 - 완주하지 못한 선수 / 난이도: Level 1
