+++ src/reeruryu/week1/P42577.java
@@ -1,4 +1,4 @@
-package reeruryu;
+package reeruryu.week1;
 
 /*
 문제: 프로그래머스 - 전화번호 목록 / 난이도: Level 2
