+++ src/reeruryu/week1/P42579.java
@@ -1,4 +1,4 @@
-package reeruryu;
+package reeruryu.week1;
 
 /*
 문제: 프로그래머스 - 베스트앨범 / 난이도: Level 3
