+++ src/reeruryu/week1/P42586.java
@@ -1,4 +1,4 @@
-package reeruryu;
+package reeruryu.week1;
 
 /*
 문제: 프로그래머스 - 기능개발 / 난이도: Level 2
