+++ src/reeruryu/week1/P42626.java
@@ -1,4 +1,4 @@
-package reeruryu;
+package reeruryu.week1;
 
 /*
 문제: 프로그래머스 - 더 맵게 / 난이도: Level 2
