+++ src/reeruryu/week1/P42627.java
@@ -1,4 +1,4 @@
-package reeruryu;
+package reeruryu.week1;
 
 /*
 문제: 프로그래머스 - 디스크 컨트롤러 / 난이도: Level 3
